fixing Install to NAND for Cubian X on CubieTruck
With latest Cubian X the install to NAND is not working anymore. It
seems that the SD card image has a different partition number for the
boot partition, therefore the sed statement does not find the correct
string to replace.
This change relaxes the search term a bit to find the boot partition on
any partition number. I hope this helps as it is untested (I did that
change myself with vi).